Immobilescout24.de scraping by submarket
This web scrapper gather all possible data (features) from immobilescout24.de for each selected district (let's call is as 'submarket') that is defined in external xls file and save it to separate Excel file with unique timestamp. There are also data visualzation part in the code that can be switchen on/off.